backend/requirements.txt
Django>=4.2
psycopg2-binary>=2.9
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
python-docx>=1.1
docxtpl>=0.16.4
django-cors-headers>=4.0.0
Pillow>=9.5.0 